Module Conex_diff_provider

Data provider using an existing provider and a diff

val apply : Conex_io.t -> Patch.t list -> Conex_io.t

val apply_diff : Conex_io.t -> string -> Conex_io.t * Patch.t list

val ids : string -> Conex_utils.path -> Patch.t list -> (bool * Conex_utils.S.t, string) Stdlib.result

ids rootname keydir diffs returns whether the root file was changed, and the set of modified ids.