Module Conex_opam_encoding

Opam file encoding

Persistent files of the opam repository should be kept in the same file format as opam files. This module uses the opam-file-format package to decode and encode the Wire.t representation into opam files.

val decode : string -> (Conex_resource.Wire.t, string) Stdlib.result

decode str is either Ok t or Error str, the input is a string in opam format.

val encode : Conex_resource.Wire.t -> string

encode t encodes t into a string in opam format.