Parameter Make.F

val ids : unit -> Conex_resource.identifier list

ids () is the list of available identifiers.

val read : (float -> Conex_resource.timestamp option) -> Conex_resource.identifier -> (string * Conex_resource.timestamp, string) Stdlib.result

read id is either the content and creation timestamp of id, or an error.

val write : Conex_resource.identifier -> string -> (unit, string) Stdlib.result

write id data stores data as id persistently, or errors.