Module Conex_resource.Delegation

type t = {

paths : Conex_utils.path list;

valid : Expression.t;

terminating : bool;

}

val equal : t -> t -> bool

equal a b is true if delegations a and b are identical.

val pp : t Conex_utils.fmt

pp is a pretty printer for a delegation.

val of_wire : Wire.s -> (t, string) Stdlib.result

of_wire w converts w to a delegation or error.

val wire_raw : t -> Wire.s

wire_raw t is the raw wire representation of t.