Module Conex_resource.Header

Common header on disk

type t = {

version : int;

created : timestamp;

counter : Conex_utils.Uint.t;

epoch : Conex_utils.Uint.t;

name : name;

typ : typ;

}

The header consists of version, created, counter, epoch, name, and typ.

val pp : t Conex_utils.fmt

pp is a pretty printer.

val wire : t -> Wire.t

wire t is the wire representation of t.

val of_wire : Wire.t -> (t, string) Stdlib.result

of_wire t converts t into a Header.t or error.

val counter : Conex_utils.Uint.t -> Conex_utils.Uint.t -> string

counter ctr epoch prints ctr, epoch to a string containing the counter and epoch (unless zero).