Module Conex_resource.Key

type alg = [

| `RSA

| `Ed25519

]

The sum type of supported asymmetric key algorithms.

val alg_to_string : alg -> string

alg_to_string a is a string for alg.

type t = identifier * timestamp * alg * string

The type of public keys

val equal : t -> t -> bool

equal a b is true if id, created, alg, and key are identical.

val pp : t Conex_utils.fmt

pp is a pretty printer for public keys

val of_wire : Wire.s -> (t, err) Stdlib.result

of_wire w converts w to a key or error.

val many_of_wire : Wire.s list -> (t Conex_utils.M.t * string list, string) Stdlib.result

many_of_wire w uses of_wire on the Wire.s.List w. Prints errors to stdout if an unknown key algorithm or multiple keys with the same identifier are used.

val wire_raw : t -> Wire.s

wire_raw t is the raw wire representation of t.

val wire : t -> Wire.t

wire t is the wire representation of t.

val keyid : (string -> Digest.t) -> t -> Digest.t

keyid hash key is hash (to_string key), the key hash.

val to_string : t -> string