Module Conex_resource.Root

The root contains the (offline) root keys, also defines snapshot, timestamp, and maintainers. Furthermore, it contains configuration information such as where keys are located in this repository and where the data is stored.

type role = [

| `Snapshot

| `Timestamp

| `Maintainer

]

val role_to_string : role -> string

module RM : sig ... end

type t = {

created : timestamp;

counter : Conex_utils.Uint.t;

epoch : Conex_utils.Uint.t;

name : identifier;

datadir : Conex_utils.path;

keydir : Conex_utils.path;

keys : Key.t Conex_utils.M.t;

valid : Expression.t;

roles : Expression.t RM.t;

signatures : Signature.t Conex_utils.M.t;

}

val t : ?counter:Conex_utils.Uint.t -> ?epoch:Conex_utils.Uint.t -> ?name:identifier -> ?datadir:Conex_utils.path -> ?keydir:Conex_utils.path -> ?keys:Key.t Conex_utils.M.t -> ?roles:Expression.t RM.t -> ?signatures:Signature.t Conex_utils.M.t -> timestamp -> Expression.t -> t

val add_signature : t -> identifier -> Signature.t -> t

val pp : t Conex_utils.fmt

val of_wire : Wire.t -> (t * string list, string) Stdlib.result

val wire : t -> Wire.t

val wire_raw : t -> Wire.t