Module Conex_resource.Snapshot

The snapshot is signed by the snapshot role. It lists all targets, with size, filename, and digest. The purpose is to prevent mix-and-match attacks, where an attacker mixes older targets files with newer ones.

type t = {

created : timestamp;

counter : Conex_utils.Uint.t;

epoch : Conex_utils.Uint.t;

name : identifier;

keys : Key.t Conex_utils.M.t;

targets : Target.t list;

signatures : Signature.t Conex_utils.M.t;

}

val t : ?counter:Conex_utils.Uint.t -> ?epoch:Conex_utils.Uint.t -> ?keys:Key.t Conex_utils.M.t -> ?targets:Target.t list -> ?signatures:Signature.t Conex_utils.M.t -> timestamp -> identifier -> t

val equal : t -> t -> bool

equal a b is true if all fields of a and b are equal.

val add_signature : t -> identifier -> Signature.t -> t

val pp : t Conex_utils.fmt

val of_wire : Wire.t -> (t * string list, string) Stdlib.result

val wire : t -> Wire.t

val wire_raw : t -> Wire.t