Module Conex_resource.Target

The target is a triple of filename, digest, and size. It is used by snapshot and targets.

type t = {

filename : Conex_utils.path;

digest : Digest.t list;

size : Conex_utils.Uint.t;

}

val equal : t -> t -> bool

equal a b is true if targets a and b are identical.

val valid_opam_path : Conex_utils.path -> t -> bool

valid_opam_path datadir t is true if the filename sticks to opam repository rules: <datadir>/foo/foo.version/opam, <datadir>/foo/foo.version/files/, <datadir>/foo.version/opam, or <datadir>/foo.version/files/.

val collect_opam_file : Conex_utils.path -> t -> bool

collect_opam_file datadir t is true if the filename sticks to opam repository rules: <datadir>/foo/foo.version/opam or <datadir>/foo.version/opam.

val pp : t Conex_utils.fmt

pp is a pretty printer for a target.

val of_wire : Wire.s -> (t, string) Stdlib.result

of_wire w converts w to a target or error.

val wire_raw : t -> Wire.s

wire_raw t is the raw wire representation of t.