Module Conex_resource.Wire

The wire encoding is abstract here, one suitable decoding and encoding engine is Conex_opam_encoding. The wire encoding is used for digest computations, and persistent storage on disk.

type s =

| Map of s Conex_utils.M.t

| List of s list

| Identifier of identifier

| Data of string

| Bigint of Conex_utils.Uint.t

| Smallint of int

| Pair of s * s

| And of s * s

| Or of s * s

The values in the key value store: either a map, a list, an identifier, data (represented as string), or an unsigned integer.

type t = s Conex_utils.M.t

The toplevel node, a Map

val to_string : t -> string

to_string t is a string representing t. This is used by Conex_verify.S to compute digests and signatures. There is no parser for this string encoding available.