Module Conex_unix_persistency

File system operations using Unix

val exists : string -> bool

exists is Sys.exists.

val mkdir : ?mode:int -> string -> (unit, string) Stdlib.result

mkdir ~mode name creates a directory name, or errors.

val remove : string -> (unit, string) Stdlib.result

remove is Sys.remove or error.

val rename : string -> string -> (unit, string) Stdlib.result

rename is Sys.rename or error.

val file_type : string -> (Conex_utils.file_type, string) Stdlib.result

file_type path is the file type of path or error.

val read_file : string -> (string, string) Stdlib.result

read_file path is the contents of path or error.

val write_file : ?mode:int -> string -> string -> (unit, string) Stdlib.result

write_file ~mode path data writes data under path or error.

val write_replace : ?mode:int -> string -> string -> (unit, string) Stdlib.result

write_replace ~mode path data writes data under path.tmp and renames path.tmp to path or error.

val collect_dir : string -> (string list, string) Stdlib.result

collect_dir path are the inhabitants of path or error.