Module type Conex_utils.LOGS

LOGS is a subset of the Logs library, providing four log streams.

module Tag : sig ... end

Tag, as in logs

type ('a, 'b) msgf = (?header:string -> ?tags:Tag.set -> ('a, Stdlib.Format.formatter, unit, 'b) Stdlib.format4 -> 'a) -> 'b

msgf

type 'a log = ('a, unit) msgf -> unit

type src

val warn_count : unit -> int

warn_count () is the number of warning messages.

val debug : ?src:src -> 'a log

debug k logs k, a to the debug stream

val info : ?src:src -> 'a log

info k logs k, a to the information stream

val warn : ?src:src -> 'a log

warn k logs k, a to the warning stream