Module type Conex_verify.S

The verification module type

val raw_digest : string -> Conex_resource.Digest.t

raw_digest str is the digest of the given str.

val digest : Conex_resource.Wire.t -> Conex_resource.Digest.t

digest wire is the digest of the string encoding of wire.

val verify : Conex_resource.Wire.t -> Conex_resource.Key.t Conex_utils.M.t -> Conex_resource.Signature.t Conex_utils.M.t -> Conex_resource.identifier Conex_resource.Digest_map.t * error list

verify wire keys sigs is the set of valid signatures and errors