conex index

Library conex

This library exposes the following toplevel modules:

Conex Establish trust in community repositories

Conex_diff_provider Data provider using an existing provider and a diff

Conex_io IO operations

Conex_opam_encoding Opam file encoding

Conex_private Private key operations and handling

Conex_repository

Conex_resource Persistent data: on wire and record types

Conex_utils String, unsigned integers, logging, collections, and more

Conex_verify Verification primitives

Library conex.openssl

The entry point of this library is the module: Conex_openssl.

Library conex.unix

This library exposes the following toplevel modules:

Conex_unix_persistency File system operations using Unix

Conex_unix_private_key Private key IO on Unix

Conex_unix_provider Unix IO provider